+++ designpattern-adapter/src/main/java/com/github/romankh3/adapter/SquarePegAdapter.java
@@ -1,0 +1,23 @@
+package com.github.romankh3.adapter;
+
+import com.github.romankh3.adapter.round.RoundPeg;
+import com.github.romankh3.adapter.square.SquarePeg;
+
+/**
+ * Adapter allows fitting square pegs into round holes.
+ */
+public class SquarePegAdapter extends RoundPeg {
+    private SquarePeg peg;
+
+    public SquarePegAdapter(SquarePeg peg) {
+        this.peg = peg;
+    }
+
+    @Override
+    public double getRadius() {
+        double result;
+        // Calculate a minimum circle radius, which can fit this peg.
+        result = (Math.sqrt(Math.pow((peg.getWidth() / 2), 2) * 2));
+        return result;
+    }
+}