+++ designpattern-adapter/src/main/java/com/github/romankh3/adapter/round/RoundHole.java
@@ -1,0 +1,28 @@
+package com.github.romankh3.adapter.round;
+
+/**
+ * RoundHoles are compatible with RoundPegs.
+ */
+public class RoundHole {
+    private double radius;
+
+    public RoundHole(double radius) {
+        this.radius = radius;
+    }
+
+    public double getRadius() {
+        return radius;
+    }
+
+    /**
+     * Show can be plug in {@link RoundPeg} into {@link RoundHole}.
+     *
+     * @param peg {@link RoundPeg} object to be plugin.
+     * @return boolean value.
+     */
+    public boolean fits(RoundPeg peg) {
+        boolean result;
+        result = (this.getRadius() >= peg.getRadius());
+        return result;
+    }
+}